requirements.txt
torch
transformers
Pillow
bentoml>=1.2.0a0
